feat(issues): Redesign streamline event nav
